move summaryMetricToDoubleGaugeMetric to StackdriverExportUtils
